refactor: replace `Watch<Option<T>>` with custom oneshot channel for signaling
Previous implementation was using `watch` channel in pair with `send_once()` and `recv_init()` methods, which was unreadable and error-prone.

It was a developer responsibility to make sure the `watch` channel is used only in pair with `send_once()` and `recv_init()` methods, which won't lead to a good outcome.

So, new custom `mpmc` `oneshot` has been implemented to become an intuitive and easy-to-use alternative.

All `send_once(), recv_init()` usages were replaced with the new channel.